add range partition support
