+++ pkg/partitionservice/service_range_test.go
@@ -1,0 +1,91 @@
+// Copyright 2021-2024 Matrix Origin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package partitionservice
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/matrixorigin/matrixone/pkg/container/types"
+	"github.com/matrixorigin/matrixone/pkg/sql/parsers/tree"
+	"github.com/matrixorigin/matrixone/pkg/txn/client"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestCreateRange(t *testing.T) {
+	num := uint64(2)
+	tableID := uint64(1)
+	columns := []string{"a"}
+	allowedT := []types.T{
+		types.T_int8,
+		types.T_int16,
+		types.T_int32,
+		types.T_int64,
+		types.T_uint8,
+		types.T_uint16,
+		types.T_uint32,
+		types.T_uint64,
+	}
+
+	for _, v := range allowedT {
+		runTestPartitionServiceTest(
+			func(
+				ctx context.Context,
+				txnOp client.TxnOperator,
+				s *service,
+				store *memStorage,
+			) {
+				def := newTestTableDefine(1, columns, []types.T{v})
+				store.addUncommittedTable(def)
+
+				stmt := newTestRangeOption(t, columns[0], num)
+				assert.NoError(t, s.Create(ctx, tableID, stmt, txnOp))
+
+				v, ok := store.uncommitted[tableID]
+				assert.True(t, ok)
+				assert.Equal(t, columns[0], v.metadata.Description)
+				assert.Equal(t, 2, len(v.partitions))
+				for _, p := range v.partitions {
+					assert.NotEqual(t, 0, p.PartitionID)
+				}
+			},
+		)
+	}
+}
+
+func newTestRangeOption(
+	t *testing.T,
+	column string,
+	num uint64,
+) *tree.CreateTable {
+	partitions := ""
+	for i := uint64(1); i <= num; i++ {
+		partitions += fmt.Sprintf("partition p%d values less than (%d)", i, i)
+		if i != num {
+			partitions += ", "
+		}
+	}
+
+	return getCreateTableStatement(
+		t,
+		fmt.Sprintf(
+			"create table t(%s int) partition by range (%s) (%s)",
+			column,
+			column,
+			partitions,
+		),
+	)
+}
