+++ pkg/partitionservice/service_test.go
@@ -16,7 +16,6 @@ package partitionservice
 
 import (
 	"context"
-	"fmt"
 	"testing"
 	"time"
 
@@ -31,47 +30,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestCreateHash(t *testing.T) {
-	num := uint64(2)
-	tableID := uint64(1)
-	columns := []string{"a"}
-	allowedT := []types.T{
-		types.T_int8,
-		types.T_int16,
-		types.T_int32,
-		types.T_int64,
-		types.T_uint8,
-		types.T_uint16,
-		types.T_uint32,
-		types.T_uint64,
-	}
-
-	for _, v := range allowedT {
-		runTestPartitionServiceTest(
-			func(
-				ctx context.Context,
-				txnOp client.TxnOperator,
-				s *service,
-				store *memStorage,
-			) {
-				def := newTestTableDefine(1, columns, []types.T{v})
-				store.addUncommittedTable(def)
-
-				stmt := newTestHashOption(t, columns[0], num)
-				assert.NoError(t, s.Create(ctx, tableID, stmt, txnOp))
-
-				v, ok := store.uncommitted[tableID]
-				assert.True(t, ok)
-				assert.Equal(t, columns[0], v.metadata.Description)
-				assert.Equal(t, 2, len(v.partitions))
-				for _, p := range v.partitions {
-					assert.NotEqual(t, 0, p.PartitionID)
-				}
-			},
-		)
-	}
-}
 
 func TestDelete(t *testing.T) {
 	num := uint64(2)
@@ -218,22 +176,6 @@ func newTestTableDefine(
 		)
 	}
 	return def
-}
-
-func newTestHashOption(
-	t *testing.T,
-	column string,
-	num uint64,
-) *tree.CreateTable {
-	return getCreateTableStatement(
-		t,
-		fmt.Sprintf(
-			"create table t(%s int) partition by hash(%s) partitions %d",
-			column,
-			column,
-			num,
-		),
-	)
 }
 
 func getCreateTableStatement(
