+++ pkg/tests/partition/create_range_test.go
@@ -1,0 +1,104 @@
+// Copyright 2021 - 2024 Matrix Origin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package partition
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/matrixorigin/matrixone/pkg/embed"
+	"github.com/matrixorigin/matrixone/pkg/pb/partition"
+	"github.com/matrixorigin/matrixone/pkg/sql/parsers"
+	"github.com/matrixorigin/matrixone/pkg/sql/parsers/dialect"
+	"github.com/matrixorigin/matrixone/pkg/tests/testutils"
+	"github.com/matrixorigin/matrixone/pkg/util/executor"
+	"github.com/stretchr/testify/require"
+)
+
+func TestCreateAndDeleteRangeBased(t *testing.T) {
+	runPartitionClusterTest(
+		t,
+		func(c embed.Cluster) {
+			cn, err := c.GetCNService(0)
+			require.NoError(t, err)
+
+			db := testutils.GetDatabaseName(t)
+			testutils.CreateTestDatabase(t, db, cn)
+
+			testutils.ExecSQLWithReadResult(
+				t,
+				db,
+				cn,
+				func(i int, s string, r executor.Result) {
+
+				},
+				fmt.Sprintf("create table %s (c int comment 'abc') partition by range (c) (partition p1 values less than (1) engine = innodb, partition p2 values less than (2) engine = innodb)", t.Name()),
+			)
+
+			metadata := getMetadata(
+				t,
+				0,
+				db,
+				t.Name(),
+				cn,
+			)
+			require.Equal(t, 2, len(metadata.Partitions))
+			require.Equal(t, partition.PartitionMethod_Range, metadata.Method)
+
+			var tables []string
+			for idx, p := range metadata.Partitions {
+				tables = append(tables, p.PartitionTableName)
+				require.NotEqual(t, uint64(0), p.PartitionID)
+				require.Equal(t, metadata.TableID, p.PrimaryTableID)
+				require.Equal(t, uint32(idx), p.Position)
+				require.Equal(t, fmt.Sprintf("%s_%s", metadata.TableName, p.Name), p.PartitionTableName)
+				require.Equal(t, fmt.Sprintf("values less than (%d)", idx+1), p.Comment)
+			}
+
+			testutils.ExecSQL(
+				t,
+				db,
+				cn,
+				fmt.Sprintf("drop table %s", t.Name()),
+			)
+			metadata = getMetadata(
+				t,
+				0,
+				db,
+				t.Name(),
+				cn,
+			)
+			require.Equal(t, partition.PartitionMetadata{}, metadata)
+
+			for _, name := range tables {
+				require.False(t, testutils.TableExists(t, db, name, cn))
+			}
+
+		},
+	)
+}
+
+func TestParse(t *testing.T) {
+	sql := `
+	create table t (
+		c DATE comment 'abc'
+	) partition by range (YEAR(c)) (
+	 	partition p1 values less than (10), 
+		partition p2 values less than (100)
+	)`
+	_, err := parsers.Parse(context.TODO(), dialect.MYSQL, sql, 0)
+	require.NoError(t, err)
+}
